Update app.js
Added additional if(!win) statement to the one(), two(), three(), four() functions. Before once I won the game the last clicked corner would still take the value of the 4 functions and override winColor(). Now once the game won all four corners are turned into similar color as it is supposed to be